requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
xlrd==2.0.1
PyPDF2==3.0.1
redis==5.0.1
aiofiles==23.2.1
pdf2image==1.16.3
Pillow==10.1.0
pytesseract==0.3.10
numpy==1.26.2
opencv-python==4.9.0.80
python-dotenv==1.0.0

# Test dependencies
pandas==2.2.3
xlwt==1.3.0
openpyxl==3.1.5